ARROW-9497: [C++][Parquet] Fix fuzz failure case caused by malformed Parquet data
I wasn't sure about the best way to fix this, but the root cause was bad data returned by `RleDecoder`. It seems that even with `bit_width_` is 1 that this class can return values greater than 1. We might have a closer look at that. I also opened ARROW-9498 about seeing if we want to make an improvement to the implementation given improvements made in Impala (where this code originated)

Closes #7779 from wesm/parquet-fuzz-fix

Authored-by: Wes McKinney <wesm@apache.org>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>