ARROW-9575: [R] gcc-UBSAN failure on CRAN
The cause of the failure itself turns out to be in cpp/src/parquet, not the R bindings.

This patch reworks the existing r-sanitizer nightly job to (1) build the bundled C++ build like what happens on CRAN, and (2) actually fail the build if there is a UBSAN error. Previously no UBSAN error was reported because the Arrow C++ library was not built with sanitizers, and even if it were, the build was not set up to detect them and fail.

Closes #7858 from nealrichardson/r-ubsan

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>