ARROW-9768 [Rust] [DataFusion] Rename PhysicalPlannerImpl to DefaultPhysicalPlanner
Proposed follow up to  https://github.com/apache/arrow/pull/7975 -- rename `PhysicalPlannerImpl` to `DefaultPhysicalPlanner` to better describe what it is and that the design allows for more than one.

I considered a more specific name like `SingleNodeMultiThreadedPlanner` but I felt a lot of that was a function of the Physical nodes that got created rather than the planner itself.

Closes #7980 from alamb/alamb/arrow-9758-rename

Authored-by: alamb <andrew@nerdnetworks.org>
Signed-off-by: Andy Grove <andygrove73@gmail.com>