ARROW-9788: [Rust] [DataFusion] Rename SelectionExec to FilterExec
The use of the term "selection" rather than "filter" was the main inconsistency that I am seeing in our APIs and was also different from the terminology used in Spark and Pandas.

I had used the term "selection" based on Codd's relational algebra [1] but it does seem that the term "filter" is more widely used so I renamed based on that.

Other changes in this PR:

- `FilterExec` now validates the filter expression to ensure it returns a boolean type.
- Made some improvements to `BinaryExpr::data_type`, although it is still not correct for the math expressions. I would prefer us to have specific math physical expressions anyway, rather than this generic 'BinaryExpr'.

[1] https://en.wikipedia.org/wiki/Relational_algebra

Closes #7998 from andygrove/ARROW-9788

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>