ARROW-10084: [Rust] [DataFusion] Added length of LargeStringArray and fixed undefined behavior.
This PR adds length to `LargeStringArray` and fixes undefined behavior.

The undefined behavior was introduced in #7878. Specifically, we were (unsafely) interpreting a `i32` buffer as `u32`.

This PR also changes the length kernel to return signed integers instead of unsigned ones. The reason for this logical change is that our offsets are stored as `i`, not `u`, and thus the kernel follows this convention (of using signed numbers to measure lengths).

Closes #8260 from jorgecarleitao/length_i64

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>