ARROW-10238: [C#] List<Struct> is broken
When reading the flatbuffer type information, we are reading the ListType incorrectly. We should be using the childFields parameter, like we do for StructTypes.

I also took this opportunity to redo how we compare schema information in our tests. For nested types, we need to recursively compare types all the way down.

@pgovind @HashidaTKS @chutchinson

Closes #8404 from eerhardt/FixListOfStruct

Authored-by: Eric Erhardt <eric.erhardt@microsoft.com>
Signed-off-by: Eric Erhardt <eric.erhardt@microsoft.com>