ARROW-10481: [R] Bindings to add, remove, replace Table columns
This allows for R users to use either `[[` or `$` syntax for removing, adding, or replacing table columns.

If one supplies an index larger than the number of columns (e.g. `mtcars[[12]] <- vector`), the column is added to the end, the index is cast to a string (here: `"12"`) for the name. This might be better as an error instead of a silent cast for the name.

Closes #8579 from jonkeane/add-table

Authored-by: Jonathan Keane <jkeane@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>