ARROW-10330: [Rust][DataFusion] Implement NULLIF() SQL function
This PR implements the NULLIF() SQL function in DataFusion.  It is implemented as a BuiltInScalarFunction, with a boolean kernel at the core which creates a new array with a modified null bitmap from the original array, based on the result of a boolean expression.   When an input data item is equal to the right side in NULLIF(), then the item's nullity becomes set in the output array.

Closes #8688 from velvia/evan/rust-datafusion-nullif-func

Lead-authored-by: Evan Chan <evan@urbanlogiq.com>
Co-authored-by: Evan Chan <velvia@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>