ARROW-10775: [Rust][DataFusion] Use ahash in join hashmap
Hashing takes up some more time in the hash joins and is visible in profiling (about 1.31% of total instructions, including for reading the csv)

This also upgrades to ahash 0.6, this takes automatically advantage of specialization on smaller keys.

Seems to have a small win on tpch benchmark query 12: ~50-100ms on my machine.

Closes #8804 from Dandandan/hashing

Authored-by: Heres, Daniel <danielheres@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>