ARROW-10337: [C++] More liberal parsing of ISO8601 timestamps with fractional seconds
The current ISO8601 timestamp parser assumes MILLI timestamps have 3 decimal places, MICRO have 6 and NANO have 9.

This patch change the parser to accept 1 to 3 digits for MILLI, 1 to 6 digits for MICRO, and 1 to 9 digits for NANO. This allows for parsing of timestamps when e.g. a CSV file does not write timestamps with trailing zeroes. See comments in ParseTimestampISO8601 for details of format.

Closes #8692 from franksmithxyz/ARROW-10337

Lead-authored-by: Frank Smith <frankiansmith@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>